+++ base/public.jl
@@ -25,6 +25,24 @@ public
     LogRange,
     UUID,
 
+# signals
+    SIGHUP,
+    SIGINT,
+    SIGQUIT,
+    SIGKILL,
+    SIGUSR1,
+    SIGUSR2,
+    SIGPIPE,
+    SIGALRM,
+    SIGTERM,
+    SIGCHLD,
+    SIGCONT,
+    SIGWINCH,
+    SIGINFO,
+    sigrtmin,
+    sigrtmax,
+    signal_name,
+
 # Semaphores
     Semaphore,
     acquire,